Rename empty.js to emptySequence.js
