+++ src/functions/tail.js
@@ -2,7 +2,7 @@ import {ArraySequence} from "../core/arrayAsSequence";
 import {error} from "../core/error";
 import {wrap} from "../core/wrap";
 
-import {EmptySequence} from "./empty";
+import {EmptySequence} from "./emptySequence";
 
 export const TailError = error({
     summary: "Failed to get tail of sequence.",
